requirements.txt
accelerate
altgraph
annotated-types
anyio
APScheduler
attrs
bcrypt
beautifulsoup4
blinker
boltons
bracex
cachetools
certifi
cffi
charset-normalizer
clang
click
click-option-group
colorama
cryptography
distro
exceptiongroup
face
filelock
filetype
Flask
flask-cors
fsspec
glom
google-auth
google-genai
googleapis-common-protos
greenlet
h11
httpcore
httpx
httpx-sse
huggingface-hub
idna
importlib_metadata
itsdangerous
Jinja2
jproperties
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain-core
langchain-google-genai
langsmith
libclang
mariadb
markdown-it-py
MarkupSafe
mcp
mdurl
mpmath
mysql-connector-python
networkx
numpy
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-threading
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
outcome
packaging
passlib
peewee
pefile
pillow
playwright
protobuf
psutil
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic-settings
pydantic_core
pyee
pyfakefs
Pygments
pyinstaller
pyinstaller-hooks-contrib
PyJWT
PyMySQL
PySocks
pysonar
pystray
python-dotenv
python-multipart
python-owasp-zap-v2.4
pywin32
pywin32-ctypes
PyYAML
referencing
regex
requests
requests-toolbelt
responses
rich
rpds-py
rsa
ruamel.yaml
ruamel.yaml.clib
safetensors
selenium
setuptools
six
sniffio
sortedcontainers
soupsieve
SQLAlchemy
sse-starlette
starlette
sympy
tenacity
tokenizers
tomli
torch
tqdm
transformers
trio
trio-websocket
typing-inspection
typing_extensions
tzdata
tzlocal
urllib3
uuid_utils
uvicorn
waitress
wcmatch
webdriver-manager
websocket-client
websockets
Werkzeug
wrapt
wsproto
zaproxy
zipp
zstandard
